+++ CHANGELOG.md
@@ -34,6 +34,7 @@ There are a couple of optional configuration for Metro's optional dependency sup
 - **Fix**: Allow `open` members from abstract graph class superclasses to be accessors.
 - **Fix**: When detecting default function/property getter bodies in graph accessors, check for `open` modality as well.
 - **Fix**: Don't duplicate includes accessor keys across multiple parent context levels.
+- **Fix**: Fix not respecting ref counting when allocating provider fields for constructor-injected class providers. This should reduce generated graph code size quite a bit.
 
 0.6.9
 -----